pyproject.toml
[tool.poetry]
name = "pytest-selenium-ui-test"
version = "0.1.0"
description = ""
authors = ["xwmyth8023 <xwmyth8023@163.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.4"
selenium = "4.19.0"
webdriver-manager = "4.0.1"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
